book_summarizer_api/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
dj-rest-auth==7.0.1
django-allauth==65.5.0
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.9
gunicorn==21.2.0
PyMuPDF==1.25.2
pypdf==5.1.0
python-dotenv==1.0.1
whitenoise==6.6.0
drf-nested-routers==0.94.1
dj-database-url==2.1.0 